MSI-1572 and MSI-1574: Grouped product support
