--- app/code/Magento/InventoryGroupedProductAdminUi/etc/adminhtml/events.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0"?>
-<!--
-/**
- * Copyright © Magento, Inc. All rights reserved.
- * See COPYING.txt for license details.
- */
--->
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:Event/etc/events.xsd">
-    <event name="controller_action_catalog_product_save_entity_after">
-        <observer name="process_save_source_items_for_grouped" instance="Magento\InventoryGroupedProductAdminUi\Observer\ProcessSourceItemsObserver"/>
-    </event>
-</config>
