+++ app/code/Magento/InventoryGroupedProductAdminUi/etc/module.xml
@@ -6,9 +6,5 @@
  */
 -->
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:Module/etc/module.xsd">
-    <module name="Magento_InventoryGroupedProductAdminUi" setup_version="1.0.0">
-        <sequence>
-            <module name="Magento_InventoryGroupedProduct"/>
-        </sequence>
-    </module>
+    <module name="Magento_InventoryGroupedProductAdminUi" setup_version="1.0.0"/>
 </config>
